feat: 스태미나 소모
최대치 소모시 체력피해

+++ Assets/Scripts/NPC/NPC.cs
@@ -242,9 +242,9 @@ public class NPC : MonoBehaviour, IDamageable
     }
 
     // 拱府利 乔秦 贸府
-    public void TakePhysicalDamage(int damage)
+    public void TakePhysicalDamage(float damage)
     {
-        health -= damage;
+        health -= (int)damage;
         if (health <= 0)
         {
             Die();
